pyproject.toml
[tool.poetry]
name = "identity-service"
version = "0.1.0"
description = ""
authors = ["Squilix01 <Squilix01@no-reply.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.116.0"}
sqlalchemy = "^2.0.41"
asyncpg = "^0.30.0"
alembic = "^1.16.4"
psycopg2 = "^2.9.10"
pyjwt = "^2.10.1"
loguru = "^0.7.3"
authlib = {extras = ["all"], version = "^1.6.1"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
